pyproject.toml
[project]
name = "nu-esports-bot"
version = "0.1.0"
description = "Northwestern Esports Discord Bot"
requires-python = ">=3.10"
dependencies = [
    "py-cord==2.6.1",
    "pyyaml==6.0.2",
    "psycopg[binary,pool]==3.2.6",
    "pillow==10.4.0",
    "aiohttp==3.11.11",
]

[dependency-groups]
dev = [
    "black==25.9.0",
    "pre-commit==4.3.0",
]

[tool.uv]
package = false

